requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
openai==0.27.0
langchain
tiktoken==0.3.0
chromadb
sentence-transformers
pydantic==1.10.2
redis==4.3.4
sqlalchemy==2.0.11
psycopg2-binary==2.9.3
reportlab==3.6.12
fpdf==1.7.2
python-dotenv==0.21.1  # To load environment variables
gunicorn